Don't add paths to sys.path, if they already exist.
This fixes the `test_plugin.py:TestPlugin:test_install_plugin` failure.
The failure happens if the whole test suite is run; it doesn't fail
if only the single test file is run (`No module named hide_toolbar`).

At the time of the failure, `sys.path` is over 1100 items long; most of
them duplicates. This somehow interacts with Python's import logic, so
that the module can't be found.