Remember how many cards we learned today
When we close the mnemosyne application and re-open it, it won't
remember the number of the already memorized cards. In other words, it
won't warn you if you've already learned 15 or more words.

This PR aims to fix this issue and re-construct the
`self._fact_ids_memorised` from the logs.

We can fetch all the newly learned cards and all the forgotten but
learned cards today. Combining the two, we can get the number of the
learned cards today.