Fix edge case of calling reset and getting an alert
When saving the Settings dialog, the `reset` is alwas called, so it 
would trigger an alert every time when, because during the reset 
function, the `warned_about_too_many_cards` is reset to 
`warned_about_too_many_cards = False`. 

This fix initializes the `warned_about_too_many_cards` on the class 
level, and calling  `reset` won`t reset it back to `False`.

Also only warn when we've just reached the 15 new or forgotten words.