Disable continuous integration. (#120)
Appveyor:

The portable flavour of Miktex we used was discontinued. We need to
find a new way to do an unattended install of Miktex or some other
distribution.

Travis:

Travis switched to Github Apps. I got it working once after the
switch, but somehow it doesn't work any more and frankly, 'debugging'
this is very frustrating. I propose investigating to switch to Azure
pipelines. This may also allow us to get rid of appveyor and only use a
single service.