Fix libpng warnings on app startup
mnemosyne prints this to the console several times on app startup:

  libpng warning: iCCP: known incorrect sRGB profile

According to [1], this is a libpng warning about some invalid content
in the ping images. Using the suggested pngcrush command, I identified
all bad png files in mnemosyne/pyqt_ui/pixmaps/ and ran them through
'mogrify', which removed the warnings. It's unclear if there are
other side effects though

[1]: https://stackoverflow.com/questions/22745076/libpng-warning-iccp-known-incorrect-srgb-profile/43415650

Signed-off-by: Cole Robinson <crobinso@redhat.com>