Fix qprocess warnings
On the linux console, mnemosyne repeatedly prints:

  QIODevice::read (QProcess): device not open

This is due to trying to interact with the 'mplayer' process
regardless of whether it's running or not. Check the process state
before trying to interact with it.

I don't have any flashcards that use the mplayer integration, so
I didn't test that this doesn't cause any unwanted issues

Signed-off-by: Cole Robinson <crobinso@redhat.com>